Differentiate the 2 structs: pom and maven project.
