+++ cli/azd/internal/appdetect/java.go
@@ -4,14 +4,11 @@ import (
 	"context"
 	"fmt"
 	"io/fs"
-	"maps"
 	"path/filepath"
-	"slices"
 	"strings"
 )
 
 type javaDetector struct {
-	rootProjects []mavenProject
 }
 
 func (jd *javaDetector) Language() Language {
@@ -22,27 +19,17 @@ func (jd *javaDetector) DetectProject(ctx context.Context, path string, entries 
 	for _, entry := range entries {
 		if strings.ToLower(entry.Name()) == "pom.xml" {
 			pomFile := filepath.Join(path, entry.Name())
-			project, err := readMavenProject(pomFile)
+			project, err := toMavenProject(pomFile)
 			if err != nil {
 				return nil, fmt.Errorf("error reading pom.xml: %w", err)
 			}
 
-			if len(project.Modules) > 0 {
+			if len(project.pom.Modules) > 0 {
 				// This is a multi-module project, we will capture the analysis, but return nil
 				// to continue recursing
-				jd.rootProjects = append(jd.rootProjects, *project)
 				return nil, nil
 			}
 
-			var currentRoot *mavenProject
-			for _, rootProject := range jd.rootProjects {
-				// we can say that the project is in the root project if the path is under the project
-				if inRoot := strings.HasPrefix(pomFile, rootProject.path); inRoot {
-					currentRoot = &rootProject
-				}
-			}
-
-			_ = currentRoot // use currentRoot here in the analysis
 			result, err := detectDependencies(project, &Project{
 				Language:      Java,
 				Path:          path,
@@ -57,26 +44,4 @@ func (jd *javaDetector) DetectProject(ctx context.Context, path string, entries 
 	}
 
 	return nil, nil
-}
-
-func detectDependencies(mavenProject *mavenProject, project *Project) (*Project, error) {
-	databaseDepMap := map[DatabaseDep]struct{}{}
-	for _, dep := range mavenProject.Dependencies {
-		if dep.GroupId == "com.mysql" && dep.ArtifactId == "mysql-connector-j" {
-			databaseDepMap[DbMySql] = struct{}{}
-		}
-
-		if dep.GroupId == "org.postgresql" && dep.ArtifactId == "postgresql" {
-			databaseDepMap[DbPostgres] = struct{}{}
-		}
-	}
-
-	if len(databaseDepMap) > 0 {
-		project.DatabaseDeps = slices.SortedFunc(maps.Keys(databaseDepMap),
-			func(a, b DatabaseDep) int {
-				return strings.Compare(string(a), string(b))
-			})
-	}
-
-	return project, nil
 }
